generate: Added generator that replaces converttrigger and transform
We have one hard-coded value here for the service action that still
needs to be generated. I haven't yet been able to determine how it came
to exist in the first place, so until I'm holding off lest I use the
wrong inputs.

Documentation updates to follow. Now we should be able to use

    generate [path].rpmConfig

to create:
- Our loads configuration file
- An importable plist with our triggers
- Our workflows required by Dimmer triggers